Extract the import flow into a store and dialogs
MainWindow held the whole import feature — the candidate scan state, the
preview-position plumbing, and the three code-built dialogs. Move the state to
an ImportStore (candidates, preview label, candidate refresh/update) that the
event router feeds and the dialogs read, and split the UI into ImportDialog,
ImportPickerDialog, and ImportConfirmDialog. The dismiss-only error dialog and
the cover-tile builder become shared DialogPrimitives, and the signal-value
middle-truncation becomes a unit-tested TextTruncation. MainWindow keeps only
the window-level drag/drop handlers, which now delegate to the store and dialog.
