+++ bae-windows/TextTruncation.cs
@@ -1,0 +1,23 @@
+﻿namespace Bae.Windows;
+
+// Middle-truncate a value so both ends stay visible: catalog numbers and barcodes
+// differ at the end, which an end-ellipsis would hide. WinUI has no middle
+// TextTrimming, and the signal-badge value is monospace (Consolas), so a
+// character budget tracks the pixel width closely. Mirrors macOS's
+// .truncationMode(.middle) on the signal value. No WinUI or bridge types, so it
+// is unit-tested.
+public static class TextTruncation
+{
+    public static string MiddleTruncate(string value, int maxChars)
+    {
+        if (value.Length <= maxChars)
+        {
+            return value;
+        }
+
+        var keep = maxChars - 1; // one character for the ellipsis
+        var head = keep / 2;
+        var tail = keep - head;
+        return value[..head] + "…" + value[^tail..];
+    }
+}
